feat: use quotes in dispatcher

+++ poem_plugins/config/__init__.py
@@ -15,8 +15,8 @@ class VersionProviderEnum(StrEnum):
 
 @unique
 class QuotesEnum(StrEnum):
-    single = '"'
-    double = "'"
+    double = '"'
+    single = "'"
 
 
 @dataclass
